pyproject.toml
[project]
name = "tsol"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.2",
    "nupy>=0.1.1",
    "pillow>=12.1.1",
    "tqdm>=4.67.3",
]
